Move demo functions to adequate files
