+++ frontend/src/sagas/demoSaga.js
@@ -1,5 +1,10 @@
 import { call, put, takeEvery, all, fork } from "redux-saga/effects";
-import { showDemoReq, uploadDemoReq, editDemoReq } from "../api/demoRequests";
+import {
+  showDemoReq,
+  uploadDemoReq,
+  editDemoReq,
+  deleteDemoReq,
+} from "../api/demoRequests";
 import {
   showDemoRequest,
   showDemoSuccess,
@@ -10,6 +15,9 @@ import {
   editDemoRequest,
   editDemoSuccess,
   editDemoError,
+  deleteDemoRequest,
+  deleteDemoSuccess,
+  deleteDemoError,
 } from "../slices/demoSlice";
 import { showDemosRequest } from "../slices/demosSlice";
 import { snackbarNotification } from "../helpers/notifierHelper";
@@ -72,10 +80,28 @@ export function* watcherEditDemo() {
   yield takeEvery(editDemoRequest, workerEditDemo);
 }
 
+export function* workerDeleteDemo({ payload }) {
+  try {
+    const { user_id, demo_id } = payload;
+    yield call(deleteDemoReq, user_id, demo_id);
+    yield put(deleteDemoSuccess());
+    yield put(snackbarNotification("Demo deleted.", "success"));
+    yield put(showDemosRequest());
+  } catch (error) {
+    yield put(deleteDemoError());
+    yield put(snackbarNotification("Failed to delete demo.", "error"));
+  }
+}
+
+export function* watcherDeleteDemo() {
+  yield takeEvery(deleteDemoRequest, workerDeleteDemo);
+}
+
 export default function* rootSaga() {
   yield all([
     fork(watcherShowDemo),
     fork(watcherUploadDemo),
     fork(watcherEditDemo),
+    fork(watcherDeleteDemo),
   ]);
 }
